// app.client/src/api/models/ExpenseTransactionListVM.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ExpenseTransactionListVM = {
    expenseTransactionID?: string | null;
    expenseId?: string | null;
    expenseTypeId?: string | null;
    name?: string | null;
    description?: string | null;
    dateOfExpense?: string;
    dateOfPayment?: string | null;
    expenseAmount?: number;
    vendorId?: string | null;
    category?: string | null;
    categoryID?: string | null;
    isCleared?: boolean;
    isVoided?: boolean;
    createdBy?: string | null;
    createdDate?: string;
    lastModifiedBy?: string | null;
    lastModifiedDate?: string | null;
    expenseName?: string | null;
    expenseTypeName?: string | null;
    isAdvance?: boolean;
    isDeposit?: boolean;
    depositReturnedDate?: string | null;
    depositReturnNotes?: string | null;
    mediaId?: string | null;
    mediaName?: string | null;
    budgetId?: string | null;
    budgetName?: string | null;
    expenseAmountApproved?: number;
    approvalId?: string | null;
    approvalName?: string | null;
    IsApproved?: boolean
};