// app.client/src/api/models/ExpenseListVM.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ExpenseListVM = {
    expenseID?: string | null;
    expenseTypeId?: string | null;
    name?: string | null;
    description?: string | null;
    isVoided?: boolean;
    createdBy?: string | null;
    createdDate?: string;
    lastModifiedBy?: string | null;
    lastModifiedDate?: string | null;
    expenseName?: string | null;
    expenseTypeName?: string | null;
    expenseCategoryName?: string | null;
};