// app.client/src/api/models/ExpenseTypeListVM.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ExpenseTypeListVM = {
    expenseTypeID?: string | null;
    name?: string | null;
    isVoided?: boolean;
    createdBy?: string | null;
    createdDate?: string;
    lastModifiedBy?: string | null;
    lastModifiedDate?: string | null;
    expenseCategoryId?: string | null;
    expenseCategoryName?: string | null;
};