// app.client/src/api/models/BudgetListVM.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type BudgetListVM = {
    budgetId?: string | null;
    name?: string | null;
    description?: string | null;
    amount?: number;
    expenseCategoryId?: string | null;
    category?: string | null;
    categoryId?: string | null;
    isVoided?: boolean;
    createdBy?: string | null;
    createdDate?: string;
    lastModifiedBy?: string | null;
    lastModifiedDate?: string | null;
    expenseCategoryName?: string | null;
    expenseTransactionTotalApproved?: number;
    expenseTransactionTotalApprovedNotPaid?: number;
    expenseTransactionTotalPaid?: number;
};